requirements.txt
streamlit>=1.35
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
groq>=0.9
pymupdf>=1.24
Pillow>=10.0
duckduckgo_search>=6.0
beautifulsoup4>=4.12
